bodh-ai-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.6
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.20